cesspool-backend/requirements.txt
django
djangorestframework
django-cors-headers
redgreenunittest
paho-mqtt